Move broken or commented-out code to attic
